requirements.txt
# GPU-enabled

# Core ML and Search (with GPU support)
sentence-transformers>=5.0.0
faiss-gpu>=1.11.0
torch>=2.0.0
numpy>=1.24.0

# Web requests and PDF processing
requests>=2.32.0
pdfminer.six>=20221105

# OCR support for Google Books PDFs
ocrmypdf>=15.0.0

# Installation instructions:
# For NVIDIA GPU with CUDA:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#   pip install faiss-gpu
#   pip install -r requirements-gpu.txt
#
# For AMD GPU with ROCm:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/rocm6.2
#   pip install faiss-gpu
#   pip install -r requirements-gpu.txt